12/4/23: call openAI API behind interaction
I didn't like the thought of putting too much logic in the QuotesController
when making the call to the openAI API, so I decided to extract the logic
and place it inside an ActiveInteraction. This makes the controller look
cleaner and do its job of focusing on handling requests.
