+++ app/controllers/quotes_controller.rb
@@ -51,38 +51,18 @@ class QuotesController < ApplicationController
   end
 
   def find_author
-    # TODO: rescue Faraday::TimeoutError, or Net::ReadTimeout when API fails
-    response =
-      client.chat(
-        parameters: {
-          model: "gpt-3.5-turbo",
-          messages: [
-            {
-              role: "user",
-              content: @quote.name
-            },
-            {
-              role: "system",
-              content: "You will be directly given a quote by the user. The
-                quote may be exactly worded correctly, or it may be paraphrased
-                incorrectly. Do your best to guess who is the author of the quote.
-                You should start off by saying 'It sounds like you may be quoting'.
-                Then, give a fun fact about the author of the quote, no more than
-                20 words. If the author of the quote is a character, then give a
-                fun fact about the character itself, not the author who created
-                the character."
-            }
-          ],
-          temperature: 0.7
-        }
-      )
+    outcome = Quotes::FindAuthor.run(quote: @quote)
 
-    @ai_response = response.dig("choices", 0, "message", "content")
+    if outcome.valid?
+      @ai_response = outcome.result
 
-    respond_to do |format|
-      format.html do
-        render "quotes/find_author", formats: [:turbo_stream]
+      respond_to do |format|
+        format.html do
+          render "quotes/find_author", formats: [:turbo_stream]
+        end
       end
+    else
+      render :edit, status: :unprocessable_entity
     end
   end
